+++ examples/auth.rs
@@ -5,47 +5,64 @@ use near_jsonrpc_client::errors::{
 use near_jsonrpc_client::{auth, methods, JsonRpcClient};
 use near_primitives::types::{BlockReference, Finality};
 
+mod utils;
+
 async fn unauthorized() -> Result<(), Box<dyn std::error::Error>> {
-    let client = JsonRpcClient::connect("https://rpc.testnet.near.org");
+    let client = JsonRpcClient::connect("https://near-mainnet.api.pagoda.co/rpc/v1/");
 
     let request = methods::block::RpcBlockRequest {
         block_reference: BlockReference::Finality(Finality::Final),
     };
 
-    let response = client.call(request).await;
-
-    assert!(
-        matches!(
-            response,
-            Err(ServerError(ResponseStatusError(Unauthorized)))
-        ),
-        "got {:?}",
-        response
-    );
+    match client.call(request).await {
+        Ok(_) => panic!("The unauthorized request succeeded unexpectedly."),
+        Err(ServerError(ResponseStatusError(Unauthorized))) => {
+            eprintln!("\x1b[33mThe unauthorized request failed as expected.\x1b[0m");
+        }
+        Err(error) => {
+            eprintln!("\x1b[31mThe unauthorized request failed with an unexpected error.\x1b[0m");
+            eprintln!("Error: {:#?}", error);
+        }
+    }
 
     Ok(())
 }
 
-async fn authorized() -> Result<(), Box<dyn std::error::Error>> {
-    let client = JsonRpcClient::connect("https://rpc.testnet.near.org")
-        .header(auth::ApiKey::new("399ba741-e939-4ffa-8c3c-306ec36fa8de")?);
+async fn authorized(api_key: &str) -> Result<(), Box<dyn std::error::Error>> {
+    let client = JsonRpcClient::connect("https://near-mainnet.api.pagoda.co/rpc/v1/")
+        .header(auth::ApiKey::new(api_key)?);
 
     let request = methods::block::RpcBlockRequest {
         block_reference: BlockReference::Finality(Finality::Final),
     };
 
-    let response = client.call(request).await?;
-
-    println!("{:#?}", response);
+    match client.call(request).await {
+        Ok(block) => println!("{:#?}", block),
+        Err(error) => {
+            eprintln!(
+                "\x1b[31mThe authorized request failed unexpectedly, is the API key valid?\x1b[0m"
+            );
+            match error {
+                ServerError(ResponseStatusError(Unauthorized)) => {
+                    println!("Unauthorized: {}", error)
+                }
+                _ => println!("Unexpected error: {}", error),
+            }
+        }
+    }
 
     Ok(())
 }
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    unauthorized().await?;
+    env_logger::init();
 
-    authorized().await?;
+    let input = utils::input("Enter an API Key: ")?;
+
+    authorized(&input).await?;
+
+    unauthorized().await?;
 
     Ok(())
 }