+++ src/methods/block.rs
@@ -23,7 +23,7 @@
 //!
 //!       ```
 //!       # use near_jsonrpc_client::methods;
-//!       # fn main() -> Result<(), Box<dyn std::error::Error>> {
+//!       # fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
 //!       use near_primitives::types::{BlockReference, BlockId};
 //!
 //!       let request = methods::block::RpcBlockRequest {
@@ -72,7 +72,7 @@ pub type RpcBlockResponse = near_primitives::views::BlockView;
 impl RpcHandlerResponse for RpcBlockResponse {}
 
 impl RpcHandlerError for RpcBlockError {
-    fn parse_raw_error(value: serde_json::Value) -> Option<Result<Self, serde_json::Error>> {
+    fn parse(value: serde_json::Value) -> Result<Self, serde_json::Error> {
         common::parse_unknown_block!(value => Self)
     }
 }